refactor: relocate changed-since git file scoping into homeboy-engine-primitives
get_files_changed_since (and its private helpers: shallow-clone ancestry
deepening, working-tree/untracked collection, diff-output parsing, default-remote
resolution) is a std-only git primitive — raw git subprocess + homeboy_error,
zero homeboy-core deps. It's shared by the audit engine, the refactor planner,
and the extension lint/test scopers. Move it from homeboy-core's git::changes
into homeboy-engine-primitives::git_changes so those consumers — including a
future homeboy-code-audit crate — depend on the slim primitives base rather than
all of homeboy-core for changed-file scoping.

- The 6-fn cluster (get_files_changed_since + get_working_tree_files_for_changed_
  since + is_shallow_repo + has_merge_base + ensure_ancestry_for_ref +
  parse_diff_output) had zero external users beyond the public entry point; it
  moves wholesale. The one super::resolve_default_remote call is inlined (trivial
  origin/sole-remote resolution) to sever the last core tie.
- ensure_ancestry_for_ref is pub so core's resolve_merge_base (which stays in
  core) reuses the same shallow-clone deepening strategy.
- homeboy-core re-exports get_files_changed_since so all 8 existing consumers
  (git::get_files_changed_since) are unaffected. The behavior test moves with the
  code; trivial smoke tests stay in core to exercise the re-export.

Extraction-prep for #8425: peels code_audit's crate::git downward dep onto the
primitives base.

Refs #8425
