+++ crates/homeboy-engine-primitives/src/lib.rs
@@ -12,6 +12,7 @@ pub mod command;
 pub mod detail_output;
 pub mod edit_op;
 pub mod edit_op_apply;
+pub mod git_changes;
 pub mod grammar;
 pub mod identifier;
 pub mod language;
